Refactor: moved user auth info to context

+++ src/utils/firebase/deleteFile.ts
@@ -3,9 +3,16 @@ import { User } from 'firebase/auth';
 import { SERVER_URL } from '../misc/constants';
 
 export async function deleteFile(
-    user: User,
+    user: User | null | undefined,
     storagePath: string
 ): Promise<ResponseStatus> {
+    if (!user) {
+        return {
+            success: false,
+            message: 'User authentication failed.',
+        };
+    }
+
     const idToken = await user.getIdToken();
     const res = await axios.delete(`${SERVER_URL}/deleteFile`, {
         headers: {
