+++ src/utils/firebase/deletePost.ts
@@ -2,7 +2,11 @@ import axios from 'axios';
 import { User } from 'firebase/auth';
 import { SERVER_URL } from '../misc/constants';
 
-export async function deletePost(user: User, postId: string) {
+export async function deletePost(
+    user: User | null | undefined,
+    postId: string
+) {
+    if (!user) return;
     const idToken = await user.getIdToken();
     return await axios.delete(`${SERVER_URL}/deletePost`, {
         headers: {
